Move all verification styles to the core
