+++ front/app/components/SignUpIn/SignUp/VerificationSignUpStep/styles.ts
@@ -1,3 +1,5 @@
+import { Label, Image } from '@citizenlab/cl2-component-library';
+import Button from 'components/UI/Button';
 import styled from 'styled-components';
 import { fontSizes, media } from 'utils/styleUtils';
 
@@ -30,4 +32,53 @@ export const Subtitle = styled.h2`
   font-size: ${fontSizes.l}px;
   font-weight: 600;
   line-height: normal;
+`;
+
+export const FormContainer = styled.div<{ inModal: boolean }>`
+  display: flex;
+  flex-direction: column;
+  align-items: ${(props) => (props.inModal ? 'center' : 'stretch')};
+`;
+
+export const Form = styled.form<{ inModal: boolean }>`
+  width: 100%;
+  max-width: ${(props) => (props.inModal ? '380px' : 'unset')};
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  margin-top: 10px;
+`;
+
+export const FormField = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  margin-bottom: 20px;
+`;
+
+export const StyledLabel = styled(Label)`
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+`;
+
+export const LabelTextContainer = styled.div`
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+`;
+
+export const Footer = styled.div`
+  display: flex;
+  padding-top: 10px;
+`;
+
+export const SubmitButton = styled(Button)`
+  margin-right: 10px;
+`;
+
+export const CancelButton = styled(Button)``;
+
+export const HelpImage = styled(Image)`
+  width: 100%;
 `;
