+++ front/app/modules/commercial/id_oostende_rrn/components/VerificationFormOostendeRrn.tsx
@@ -19,7 +19,7 @@ import {
   Footer,
   SubmitButton,
   CancelButton,
-} from 'modules/commercial/verification/citizen/components/styles';
+} from 'components/SignUpIn/SignUp/VerificationSignUpStep/styles';
 
 // hooks
 import useAuthUser from 'hooks/useAuthUser';
